DRY texture img selection/dropzone w/hook
