+++ src/components/TextureView.tsx
@@ -3,7 +3,11 @@ import { mdiMenuLeftOutline, mdiMenuRightOutline } from '@mdi/js';
 import Icon from '@mdi/react';
 import Img from 'next/image';
 import useViewportSizes from 'use-viewport-sizes';
-import { useObjectNavControls, useObjectUINav } from '@/hooks';
+import {
+  useObjectNavControls,
+  useObjectUINav,
+  useTextureReplaceDropzone
+} from '@/hooks';
 import {
   selectTextureIndex,
   selectUpdatedTextureDefs,
@@ -67,39 +71,14 @@ const Styled = styled('div')(
 
 export default function TextureView() {
   useObjectNavControls();
-  const dispatch = useAppDispatch();
   const uiControls = useObjectUINav();
   const [vpW] = useViewportSizes();
   const size = Math.round((vpW - 222) * 0.66);
   const textureIndex = useAppSelector(selectTextureIndex);
   const textureDefs = useAppSelector(selectUpdatedTextureDefs);
 
-  // @TODO: DRY w GuiPanelTexture
-  const onSelectNewImageFile = useCallback(
-    async (imageFile: File) => {
-      dispatch(selectReplacementTexture({ imageFile, textureIndex }));
-    },
-    [textureIndex]
-  );
-
-  const onDrop = useCallback(
-    async ([file]: File[]) => {
-      onSelectNewImageFile(file);
-    },
-    [onSelectNewImageFile]
-  );
-
-  const { getRootProps: getDragProps, isDragActive } = useDropzone({
-    onDrop,
-    multiple: false,
-    noClick: true,
-    accept: {
-      'image/bmp': ['.bmp'],
-      'image/png': ['.png'],
-      'image/jpeg': ['.jpg', '.jpeg'],
-      'image/gif': ['.gif']
-    }
-  });
+  const { isDragActive, getDragProps } =
+    useTextureReplaceDropzone(textureIndex);
 
   const textureUrl = textureDefs?.[textureIndex]?.dataUrls?.opaque;
