+++ src/hooks/index.ts
@@ -5,3 +5,4 @@ export { default as useHeldRepetitionTimer } from './useHeldRepetitionTimer';
 export { default as useDebouncedEffect } from './useDebouncedEffect';
 export { default as useSceneGLTFFileDownloader } from './useSceneGLTFFileDownloader';
 export { default as useSupportedFilePicker } from './useSupportedFilePicker';
+export { default as useTextureReplaceDropzone } from './useTextureReplaceDropzone';
